requirements.txt
python-dotenv
aiogram
openai
executor